% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_convert_mb.R
\name{get_convert_mb}
\alias{get_convert_mb}
\title{Read series from Macrobond to a data.frame.}
\usage{
get_convert_mb(series, frequency = "Highest", currency = "",
  start_date = "", end_date = "")
}
\arguments{
\item{series}{A (named) vector of  Macrobond serie codes.}

\item{frequency}{An option for frequncy conversion. Default is \code{Highest},
for others see \code{\link[MacrobondAPI]{TimeSeriesFrequency}}.}

\item{currency}{An option for currency conversion. Default is \code{""}
(no conversion), for conversions use currency codes ("USD", "EUR", etc).
List of supported currencies: \url{https://www.macrobond.com/index.php/currencylist}.}

\item{start_date, }{A starting date. A format of available absolute and
relative date formulations see \code{\link[MacrobondAPI]{SeriesRequest-class}}}

\item{end_date, }{A ending date. A format of available absolute and
relative date formulations see \code{\link[MacrobondAPI]{SeriesRequest-class}}}
}
\value{
A data.frame with dates in \code{time} column and
  time series on other (named) columns.
}
\description{
The function is used to read one or several time series from
the Macrobond. If time series have different frequency they are converted
to same frequency. Other conversions in macrobond are also available.
}
\details{
For time series of same frequnecy the \code{\link{get_mb}} is faster.
}
\examples{
  xx <- get_convert_mb(series = c("usgdp", "uscpi"))
  xx <- get_convert_mb(series = c(gdp = "usgdp", cpi = "uscpi"))
  yy <- get_convert_mb(series = c("usgdp", "uscpi"), frequency = "Annual")
 zz <- get_convert_mb(series = c("usnaac0057", "eunaac0019"),
               frequency = "Annual", currency = "USD",
               start_date = "", end_date = "")
}
