% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mb.R
\name{get_mb}
\alias{get_mb}
\title{Get series from Macrobond in a data.frame.}
\usage{
get_mb(series)
}
\arguments{
\item{series}{a (named) vector of Macrobond serie codes.}
}
\value{
A data.frame with dates in \code{time} column and
  time series on other (named) columns.
}
\description{
The function is used to get one or several time series of same frequency from
the Macrobond.
}
\details{
For time series of different frequnecies use \code{\link{get_convert_mb}}.
}
\examples{
  x <- get_mb(series = c("usgdp", "figdp"))
  x <- get_mb(series = c(USA = "usgdp", Finland = "figdp"))
}
